Update sensor loop parameter code generation format
